web/backend/requirements.txt
# VieNeu TTS Backend — Dependencies
# Note: torch is installed separately in Dockerfile with CUDA index

# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Settings
pydantic-settings==2.7.1

# Audio processing
torchaudio>=2.0.0
soundfile==0.13.1
numpy>=2.0.0
scipy>=1.10.0

# VieNeu TTS SDK
vieneu==1.3.0

# LoRA / Fine-tuning
peft>=0.15.0
transformers>=5.0.0
accelerate>=1.0.0
safetensors>=0.4.0

# HuggingFace model download
huggingface_hub>=1.0.0

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0
